Split "client" into "client-interface" with all the java code, so that "client" has nothing but typescript and sass.
