+++ client-interface/src/main/java/java2ts/Bookmark.java
