+++ client-interface/src/main/java/java2ts/ImageProps.java
@@ -1,6 +1,6 @@
 /*
  * MyTake.org website and tooling.
- * Copyright (C) 2017 MyTake.org, Inc.
+ * Copyright (C) 2017-2020 MyTake.org, Inc.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -20,9 +20,8 @@
 package java2ts;
 
 @jsweet.lang.Interface
-public class VideoBlock {
-	public StringTypes.video kind;
-	public String videoId;
-	@jsweet.lang.Optional
-	public jsweet.util.tuple.Tuple2<Integer, Integer> range;
+public class ImageProps {
+	public String dataUri;
+	public String width;
+	public String height;
 }
