+++ client-interface/src/main/java/java2ts/Json.java
