+++ client-interface/src/main/java/java2ts/LoginApi.java
