+++ client-interface/src/main/java/java2ts/LoginCookie.java
