+++ client-interface/src/main/java/java2ts/ParagraphBlock.java
@@ -1,6 +1,6 @@
 /*
  * MyTake.org website and tooling.
- * Copyright (C) 2017 MyTake.org, Inc.
+ * Copyright (C) 2017-2020 MyTake.org, Inc.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -20,7 +20,7 @@
 package java2ts;
 
 @jsweet.lang.Interface
-public class PublishResult implements Json {
-	public String publishedUrl;
-	public boolean conflict;
+public class ParagraphBlock {
+	public StringTypes.paragraph kind;
+	public String text;
 }
