+++ client-interface/src/main/java/java2ts/PublishResult.java
@@ -1,6 +1,6 @@
 /*
  * MyTake.org website and tooling.
- * Copyright (C) 2017 MyTake.org, Inc.
+ * Copyright (C) 2017-2020 MyTake.org, Inc.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License as published by
@@ -19,33 +19,8 @@
  */
 package java2ts;
 
-@jsweet.lang.Erased
-public interface StringTypes {
-	// ParagraphBlock / DocumentBlock / VideoBlock
-	@jsweet.lang.StringType
-	@jsweet.lang.Erased
-	public interface paragraph {}
-
-	@jsweet.lang.StringType
-	@jsweet.lang.Erased
-	public interface document {}
-
-	@jsweet.lang.StringType
-	@jsweet.lang.Erased
-	public interface video {}
-
-	// DocumentFact
-	@jsweet.lang.StringType
-	@jsweet.lang.Erased
-	public interface ratified {}
-
-	@jsweet.lang.StringType
-	@jsweet.lang.Erased
-	public interface published {}
-
-	// VideoFact
-	@jsweet.lang.StringType
-	@jsweet.lang.Erased
-	public interface recorded {}
-
+@jsweet.lang.Interface
+public class PublishResult implements Json {
+	public String publishedUrl;
+	public boolean conflict;
 }
